# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "UsersController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My account, roles and any consents I still need to give",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "preferredLanguage": {
                    "type": "string",
                    "enum": [
                      "en",
                      "hi",
                      "ta",
                      "kn",
                      "te"
                    ]
                  },
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "en",
                        "hi",
                        "ta",
                        "kn",
                        "te"
                      ]
                    },
                    "minItems": 1,
                    "maxItems": 5
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update my name and languages",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete_v1",
        "parameters": [],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request account deletion. Sessions end now; data is erased after a 30-day grace period.",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/roles": {
      "post": {
        "operationId": "UsersController_addRole_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "CUSTOMER",
                      "WORKER",
                      "EMPLOYER"
                    ]
                  }
                },
                "required": [
                  "role"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a role (Customer, Ranger or Employer). Student arrives with Haggler Campus.",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/consents": {
      "get": {
        "operationId": "UsersController_consents_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My consent history",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_grant_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "TERMS_OF_SERVICE",
                      "PRIVACY_POLICY",
                      "KYC_PROCESSING"
                    ]
                  },
                  "version": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  }
                },
                "required": [
                  "purpose",
                  "version"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Give consent for a purpose at the current legal version",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/consents/{purpose}": {
      "delete": {
        "operationId": "UsersController_withdraw_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw consent for a purpose",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/export": {
      "get": {
        "operationId": "UsersController_export_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download all data we hold about me (DPDP access right)",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/addresses": {
      "get": {
        "operationId": "UsersController_listAddresses_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My saved addresses",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_createAddress_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 30
                  },
                  "line1": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 120
                  },
                  "line2": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "city": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "state": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "pincode": {
                    "type": "string",
                    "pattern": "^[1-9][0-9]{5}$"
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "isDefault": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "label",
                  "line1",
                  "city",
                  "state",
                  "pincode"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save an address. Send latitude+longitude from the phone GPS, or we try to geocode it.",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/addresses/{id}": {
      "patch": {
        "operationId": "UsersController_updateAddress_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 30
                  },
                  "line1": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 120
                  },
                  "line2": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "city": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "state": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "pincode": {
                    "type": "string",
                    "pattern": "^[1-9][0-9]{5}$"
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "isDefault": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit one of my addresses",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteAddress_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one of my addresses",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/emergency-contacts": {
      "get": {
        "operationId": "UsersController_listContacts_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My emergency contacts (used by SOS and live-trip sharing)",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_createContact_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+91[6-9][0-9]{9}$"
                  },
                  "relationship": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 40
                  }
                },
                "required": [
                  "name",
                  "phone",
                  "relationship"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an emergency contact (max 5)",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/emergency-contacts/{id}": {
      "patch": {
        "operationId": "UsersController_updateContact_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+91[6-9][0-9]{9}$"
                  },
                  "relationship": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 40
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit an emergency contact",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteContact_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an emergency contact",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Haggler API",
    "description": "REST API for Haggler. Money is integer paise. Errors use the envelope { error: { code, message, details?, requestId } }. Workers are called \"Rangers\" in the app.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}